Fixed #2: Added async fetching mode
If GexfParser.parse is called with two arguments (the URL and a callback), then the file will be loaded asynchronously and the callback will be executed with the final graph as only argument at the end of the process.
If GexfParser.parse is called with only a single argument (the URL), then it works as before and it synchronously returns the final graph.
Also added a unit test checking that sync and async mode do work the same.